requirements.txt
flask-2.3.3
numpy-1.24.3
scikit-learn-1.2.2
pandas-2.10

